Using test to create a NewClient for account APIs
I'm still trying to understand how golang works for interfaces
overload of the methods and testing using mocks.
Making mistakes and progressing, creating and destroying code.
Now, I'll move on with the next task of creating the first end to end
test to create a GB account type using a template similar to the one
used in the Form3 API's documentation example